backend/requirements.txt
# Workspace version: 1.42.60
databutton==0.38.34
ipykernel==6.29.5
fastapi==0.109.2
python-multipart==0.0.9
uvicorn[standard]==0.27.1

openai
beautifulsoup4
requests

# File processing dependencies
PyPDF2==3.0.1
python-docx==1.1.2
aiohttp>=3.8.0
supabase>=1.0.0
python-dotenv==1.0.1